requirements.txt
apispec==6.3.0
async-timeout==4.0.2
blinker==1.6.2
click==8.1.6
Flask==2.3.2
flask-smorest==0.42.0
Flask-SQLAlchemy==3.0.5
Flask-WTF==1.1.1
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
packaging==23.1
psycopg2==2.9.6
python-decouple==3.8
python-dotenv==1.0.0
redis==4.6.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
webargs==8.3.0
Werkzeug==2.3.6
WTForms==3.0.1
